19/12/2022

+++ Entities/Concrete/Car.cs
@@ -8,8 +8,8 @@ namespace Entities.Concrete
     public class Car : IEntity
     {
         public int CarId { get; set; }
-        public int BrandID { get; set; }
-        public int ColorId { get; set; }
+        public short BrandID { get; set; }
+        public byte ColorId { get; set; }
         public string ModelYear { get; set; }
         public decimal DailyPrice { get; set; }
         public string Description { get; set; }
